JSON Schema refactoring, and '_metadata' added
Now validation Perl5 and Python validators are able to resolve relative
JSON Pointer references among schemas, refactor common definitions.

Added '_metadata' to all the JSON Schema elements. The objects under these
keys contain all the metadata needed in the different OpenEBench processing
and platform steps. Its contents should be only generated, validated or
managed by OpenEBench programs. The keys of the properties in this object
must start either with an underscore or with a namespace. This last case
could be considered in the future to identify accesory JSON Schemas which
help on further validation.