Implemented both participation model (issue #126) and anonymous participants (issue #125)
On one hand, the participation model at the challenge level describes
how a challenge is going to happen:

* Either publicly: results are visible ASAP, even before the challenge ends.

* Anonymous: results are visible ASAP, but you cannot know the participant details.

* Embargoed: results are only visible through the dataset visibility rules,
  and open to the public when the challenge ends.

* Secret: results are only visible anonymized through the dataset visibility rules,
  and open to the public when the challenge ends.

Then, both datasets and test actions have gained an anonymize attribute, which rule
whether the entry should be anonymized, and when the anonymization ends. The visibility
of the dataset entry is still ruled by the visibility attribute.